make live public repo

+++ app/helpers/application_helper.rb
@@ -9,16 +9,15 @@ module ApplicationHelper
   end
 
   def status_style(status)
-    color = 'text-success'
-
+    color = 'label-success'
     case status
     when 'Referred'
-      color = 'text-danger'
+      color = 'label-danger'
     when 'Investigating'
-      color = 'text-warning'
+      color = 'label-warning'
     end
 
-    content_tag(:span, class: color) do
+    content_tag(:span, class: ['label' , color]) do
       status
     end
   end
@@ -33,6 +32,26 @@ module ApplicationHelper
 
   def removeable? object, associated_objects_count
     if associated_objects_count.zero?
+      link_to object, method: 'delete', data: { confirm: 'Are you sure you want to delete?' } do
+        content_tag(:i, '', class: 'glyphicon glyphicon-trash')
+      end
+    else
+      content_tag(:i, '', class: 'glyphicon glyphicon-trash')
+    end
+  end
+
+  def domain_removeable? object, associated_objects_count
+    if associated_objects_count.zero? && object.assessment_domains.blank?
+      link_to object, method: 'delete', data: { confirm: 'Are you sure you want to delete?' } do
+        content_tag(:i, '', class: 'glyphicon glyphicon-trash')
+      end
+    else
+      content_tag(:i, '', class: 'glyphicon glyphicon-trash')
+    end
+  end
+
+  def client_removeable? object, associated_objects_count
+    if associated_objects_count[0].zero? && associated_objects_count[1].zero?
       link_to object, method: 'delete', data: { confirm: 'Are you sure you want to delete?' } do
         content_tag(:i, '', class: 'glyphicon glyphicon-trash')
       end
@@ -76,6 +95,54 @@ module ApplicationHelper
     else
       ''
     end
+  end
+
+  def user_dashboard(user)
+    user.admin? || user.any_case_manager? ? 'col-md-6' : ''
+  end
+
+  def user_dashboard_responsive(user)
+    user.admin? || user.case_worker? ? 'box-pusher' : 'col-xs-12 big-box'
+  end
+
+  def hidden_class(bool)
+    'hidden' if bool
+  end
+
+  def exit_modal_class(bool)
+    bool ? 'exitFromCif' : 'exitFromCase'
+  end
+
+  def dynamic_md_col(user)
+    user.any_case_manager? ? 'col-md-12' : 'col-md-6'
+  end
+
+  def dynamic_third_party_cols(user)
+    if user.admin?
+      'col-sm-6 col-md-3'
+    elsif user.any_case_manager?
+      'col-sm-4'
+    elsif user.able_manager? || user.case_worker?
+      'col-sm-6'
+    end
+  end
+
+  def custom_case_dashboard(user)
+    if user.any_case_manager?
+      'big-box'
+    else user.admin? || user.case_worker?
+      'top-spacing'
+    end
+  end
+
+  def error_notification(f)
+    if f.error_notification.present?
+      content_tag(:div, t('review_problem'), class: 'alert alert-danger')
+    end
+  end
+
+  def able_related_info(value)
+    'able-related-info' if ['illness', 'disability'].any? { |w| value.include?(w) }
   end
 
 end
